+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/_internal/ui/menu/ToolsMenuTree.kt
@@ -147,6 +147,9 @@ internal class ToolsMenuTree(
           onSuccess()
         }
       }
+      else {
+        throw IllegalStateException("Can't find developer tool with ID: $developerToolId")
+      }
     }
   }