+++ CHANGELOG.md
@@ -10,6 +10,21 @@
 
 ### Fixed
 
+## 4.1.0 - 2024-03-24
+
+### Added
+
+- Add automatic input text case detection to the text case converter 
+- Add escape/unescape as editor actions and code intentions
+- Add new tool: Text Statistic
+- Add support for the "Dot Text Case"
+- Add common hashing algorithms to the encoding editor action and code intention
+- Add common SHA3 algorithms to the random data generator editor action
+
+### Fixed
+
+- Editor actions on Java Strings will now preserve the outer String quotations
+
 ## 4.0.0 - 2024-03-10
 
 ### Added